+++ docs/changelog.rst
@@ -54,6 +54,8 @@ Other changes:
 
 - Deprecated Phone Number fields
   (`gh-262 <https://github.com/django/django-localflavor/pull/262>`_)
+- Bumped versions of requirements for testing.
+  (`gh-274 <https://github.com/django/django-localflavor/pull/274>`_)
 
 
 1.3   (2016-05-06)