+++ docs/changelog.rst
@@ -51,6 +51,9 @@ Other changes:
 - Ensure the migration framework generates schema migrations for model fields that change the max_length
   (`gh-257 <https://github.com/django/django-localflavor/pull/257>`_). Users will need to generate migrations for any
   model fields they use with 'makemigrations'.
+- Lazily generate US_STATES, STATE_CHOICES, and USPS_CHOICES
+  (`gh-203 <https://github.com/django/django-localflavor/issues/203>`_
+  `gh-272 <https://github.com/django/django-localflavor/pull/272>`_).
 - Deprecated Phone Number fields
   (`gh-262 <https://github.com/django/django-localflavor/pull/262>`_).
 - Bumped versions of requirements for testing